refactor(tests): split test files
