+++ test/check-requirements/check-requirements.test.ts
@@ -1,5 +1,3 @@
-import type { SemVer } from "semver";
-
 import childProcess from "node:child_process";
 import process from "node:process";
 
@@ -7,40 +5,8 @@ import semver from "semver";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { checkRequirements } from "../../src/check-requirements/index.js";
-import { isGitVersionCompatible } from "../../src/check-requirements/is-git-version-compatible.js";
-import { isNodeVersionCompatible } from "../../src/check-requirements/is-node-version-compatible.js";
 
 import { GIT_MIN_VERSION, REQUIRED_NODE_VERSIONS } from "../../src/check-requirements/constants.js";
-
-describe("incompatible Node version", () => {
-  it("tests an outdated Node version", () => {
-    expect(isNodeVersionCompatible("v16.20.2", REQUIRED_NODE_VERSIONS)).toBe(false);
-  });
-  it("tests an odd Node version", () => {
-    expect(isNodeVersionCompatible("v23.7.0", REQUIRED_NODE_VERSIONS)).toBe(false);
-  });
-});
-
-describe("compatible Node version", () => {
-  it("tests a Node maintenance LTS version", () => {
-    expect(isNodeVersionCompatible("v18.20.6", REQUIRED_NODE_VERSIONS)).toBe(true);
-  });
-  it("tests a Node active LTS version", () => {
-    expect(isNodeVersionCompatible("v20.18.2", REQUIRED_NODE_VERSIONS)).toBe(true);
-  });
-  it("tests a Node current version (if even)", () => {
-    expect(isNodeVersionCompatible("v22.13.1", REQUIRED_NODE_VERSIONS)).toBe(true);
-  });
-});
-
-describe("git version", () => {
-  it("tests an incompatible Git version", () => {
-    expect(isGitVersionCompatible("2.7.1" as unknown as SemVer)).toBe(false);
-  });
-  it("tests a compatible Git version", () => {
-    expect(isGitVersionCompatible("2.48.1" as unknown as SemVer)).toBe(true);
-  });
-});
 
 describe("check requirements", () => {
   const formerLtsReleases = [
